+++ src/main/java/uk/gov/hmcts/opal/dto/reference/CourtReferenceData.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.opal.entity.projection;
+package uk.gov.hmcts.opal.dto.reference;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
