+++ src/main/java/uk/gov/hmcts/opal/dto/reference/CourtReferenceDataResults.java
@@ -4,7 +4,6 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import uk.gov.hmcts.opal.entity.projection.CourtReferenceData;
 
 import java.util.List;
 import java.util.Optional;
