+++ src/main/java/uk/gov/hmcts/opal/entity/DefendantAccountEntity.java
@@ -27,6 +27,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import uk.gov.hmcts.opal.entity.converter.DefendantAccountTypeConverter;
+import uk.gov.hmcts.opal.entity.court.CourtEntity;
 import uk.gov.hmcts.opal.util.LocalDateAdapter;
 
 import java.math.BigDecimal;
