+++ src/main/java/uk/gov/hmcts/opal/repository/CourtRepository.java
@@ -4,7 +4,7 @@ package uk.gov.hmcts.opal.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.springframework.stereotype.Repository;
-import uk.gov.hmcts.opal.entity.CourtEntity;
+import uk.gov.hmcts.opal.entity.court.CourtEntity;
 
 @Repository
 public interface CourtRepository extends JpaRepository<CourtEntity, Long>,
