+++ src/main/java/uk/gov/hmcts/opal/repository/jpa/CourtSpecs.java
@@ -7,8 +7,8 @@ import jakarta.persistence.criteria.Predicate;
 import org.springframework.data.jpa.domain.Specification;
 import uk.gov.hmcts.opal.dto.search.CourtSearchDto;
 import uk.gov.hmcts.opal.entity.BusinessUnitEntity;
-import uk.gov.hmcts.opal.entity.CourtEntity;
-import uk.gov.hmcts.opal.entity.CourtEntity_;
+import uk.gov.hmcts.opal.entity.court.CourtEntity;
+import uk.gov.hmcts.opal.entity.court.CourtEntity_;
 import uk.gov.hmcts.opal.entity.LocalJusticeAreaEntity;
 
 import java.util.Optional;
